% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-anarchy.R
\docType{data}
\name{anarchy}
\alias{anarchy}
\title{The Mask of Anarchy}
\format{
A VCorpus of 1 text document.
}
\source{
\enumerate{
   \item Percy Bysshe Shelley, \emph{The Masque of Anarchy}, London: Edward Moxon, 1832, page 1-42. See \url{https://archive.org/stream/masqueanarchyap00huntgoog} and \url{https://knarf.english.upenn.edu/PShelley/anarchy.html}.
   \item Wikimedia Foundation, Inc. Wikipedia, 14 April 2016, “The Masque of Anarchy”, \url{https://en.wikipedia.org/wiki/The_Masque_of_Anarchy}.
}
}
\usage{
anarchy
}
\description{
A dataset which contains one poem written by Percy Bysshe Shelley that
expresses his outrage at the Peterloo Massacre of 1819. ""The Masque of
Anarchy" (or "The Mask of Anarchy") is a British political poem written in
1819 by Percy Bysshe Shelley. In his call for freedom, it is perhaps the
first modern statement of the principle of nonviolent resistance." Source:
Wikipedia.
}
\keyword{datasets}
