% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-appeal_womanhood_world.R
\docType{data}
\name{appeal_womanhood_world}
\alias{appeal_womanhood_world}
\title{Appeal to Womanhood Throughout the World}
\format{
A VCorpus of 1 text document.
}
\source{
Julia Ward Howe, "Appeal to Womanhood Throughout the World", September 1870, Boston, Massachusetts. See \url{https://www.loc.gov/resource/rbpe.07400300}.
}
\usage{
appeal_womanhood_world
}
\description{
A dataset which contains one document written by Julia Ward Howe
that asks that Christian women all over the world come together
to start a movement for global peace and unity.
}
\keyword{datasets}
