% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-connect-dots.R
\docType{data}
\name{connect_dots}
\alias{connect_dots}
\title{Connect the D.O.T.S.}
\format{
A VCorpus of 1 text document.
}
\source{
Obiora Embry, "Connect the D.O.T.S.", \emph{Expanding Horizons Through Creative Expressions: Reflections and Thoughts Related to the Struggle for Peace, Sustainability, Equality, and the Search for Humanity}, Lexington, Kentucky: Obiora Embry, 2013, page 78-82. See \url{https://www.embrybooks.org/books/expanding-horizons/}.
}
\usage{
connect_dots
}
\description{
A dataset which contains a piece written by Obiora Embry in April 2006. It
was written to help wake up consumers and empower them with knowledge to
invoke change from automobile manufacturers and within society by connecting
various dots.
}
\keyword{datasets}
