% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemisctext.R
\docType{package}
\name{iemisctext}
\alias{iemisctext}
\alias{iemisctext-package}
\title{iemisctext: Irucka Embry's miscellaneous data collection of text documents}
\description{
iemisctext provides an eclectic set of written expressions by Percy Bysshe
Shelley, Mark Twain, Major General Smedley Butler, Obiora Embry, and Irucka
Ajani Embry.
}

