% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-war-racket.R
\docType{data}
\name{war_racket}
\alias{war_racket}
\title{War Is A Racket}
\format{
A VCorpus of 1 text document.
}
\source{
Major General Smedley Butler, \emph{War Is A Racket}, New York City, New York: Round Table Press, Inc., 1935, page 1-14. See \url{https://archive.org/details/WarIsARacket} and \url{https://ratical.org/ratville/CAH/warisaracket.html}.
}
\usage{
war_racket
}
\description{
A dataset which contains one book written by Major General Smedley Butler
that describes how war is and has always been a racket.
}
\keyword{datasets}
